Refactor: Se ha editado la entidad de Usuario por problema de sintaxis

+++ src/main/java/com/deleondiego/kinalapp/entity/Usuario.java
@@ -10,7 +10,7 @@ import jakarta.persistence.Table;
 public class Usuario {
     @Id
     @Column (name = "codigo_usuario")
-    private long codigoUsuario;
+    private Long codigoUsuario;
     @Column
     private String userName;
     @Column
@@ -20,12 +20,12 @@ public class Usuario {
     @Column
     private String rol;
     @Column
-    private int estado;
+    private Long estado;
 
     public Usuario() {
     }
 
-    public Usuario(long codigoUsuario, String userName, String password, String email, String rol, int estado) {
+    public Usuario(Long codigoUsuario, String userName, String password, String email, String rol, Long estado) {
         this.codigoUsuario = codigoUsuario;
         this.userName = userName;
         this.password = password;
@@ -34,11 +34,11 @@ public class Usuario {
         this.estado = estado;
     }
 
-    public long getCodigoUsuario() {
+    public Long getCodigoUsuario() {
         return codigoUsuario;
     }
 
-    public void setCodigoUsuario(long codigoUsuario) {
+    public void setCodigoUsuario(Long codigoUsuario) {
         this.codigoUsuario = codigoUsuario;
     }
 
@@ -74,11 +74,11 @@ public class Usuario {
         this.rol = rol;
     }
 
-    public int getEstado() {
+    public Long getEstado() {
         return estado;
     }
 
-    public void setEstado(int estado) {
+    public void setEstado(Long estado) {
         this.estado = estado;
     }
 }
